param-server/requirements.txt
flask==2.0.1
werkzeug==2.0.3
pymongo==3.12.0
python-dotenv==0.19.0
flask-cors==3.0.10
requests==2.31.0
python-jose==3.3.0
tapipy==1.2.0 
pandas==2.2.3
numpy
scikit-learn
flask-socketio

torch